for #1238 Add LocalTransactionManager.
